conform with C11 and set some warning flags

+++ src/connection.h
@@ -1,7 +1,7 @@
 #ifndef CONNECTION_H
 #define CONNECTION_H
 
-int connect_ap(int sock, const char *ip, unsigned int port);
+int connect_ap(int sock, const char *ip, unsigned short port);
 int create_sock(void);
 
 #endif
